fix: rename telemetry_common to tracing_setup
